[tests] add test for class attribute parsing
[parser] add type argument parsing for method definitions

[parser] read methods from smali file

[clean init in string method]

[metaclass] embryonary support for java classes

The base idea is to use a metaclass for returning the class corresponding to the
smali code, and be able to instanciate such classes as python class and execute
seamlessly object methods and such.

pep8

[py3 compat] py3/py2 compat quirks

[emulator] various improvements

[doc] add a documentation folder

[emulator/scripting] add scripting support

[emulator] succ. method calls

Allow successive method calls in  method by:
  - resetting vm state after execution of a given method
  - passing the current vm state to the newly executed method

A cleaner design would eventually be needed ASAP but it works well in the
current state.

[doc] update doc for method calls

[doc] update README.md

[opcodes] add ushr-int opcode

[classes] add support for classes

[smaliv2] big refacto of architecture

[fix some tests]

[fix more tests]

[py2/3 compat update]